+++ lib/kion/kion.go
@@ -63,7 +63,13 @@ func runQuery(method string, url string, token string, query map[string]string, 
 	return respBody, resp.StatusCode, nil
 }
 
-// getVersion returns the targeted Kion's version number.
+////////////////////////////////////////////////////////////////////////////////
+//                                                                            //
+//  Kion Configurations                                                       //
+//                                                                            //
+////////////////////////////////////////////////////////////////////////////////
+
+// GetVersion returns the targeted Kion's version number.
 func GetVersion(host string, token string) (string, error) {
 	url := fmt.Sprintf("%v/api/version", host)
 	query := map[string]string{}
@@ -84,4 +90,35 @@ func GetVersion(host string, token string) (string, error) {
 	}
 
 	return response.Version, nil
+}
+
+// GetSessionDuration returns the AWS session duration configuration Kion uses
+// to generate session tokens. If 403 is received, we assume the shortest
+// setting of 15 minutes.
+func GetSessionDuration(host string, token string) (int, error) {
+	url := fmt.Sprintf("%v/api/v3/app-config/aws-access", host)
+	query := map[string]string{}
+	var data interface{}
+	resp, status, err := runQuery("GET", url, token, query, data)
+	if err != nil {
+		if status == 403 {
+			return 15, nil
+		} else {
+			return 0, err
+		}
+	}
+
+	// unmarshal response body
+	var response struct {
+		Status int `json:"status"`
+		Data   struct {
+			Duration int `json:"aws_temporary_credentials_duration"`
+		} `json:"data"`
+	}
+	err = json.Unmarshal(resp, &response)
+	if err != nil {
+		return 0, err
+	}
+
+	return response.Data.Duration, nil
 }